Updated hires test problems
