+++ examples/PowerElectronics/Partition/HiresBus.hpp
@@ -6,7 +6,7 @@
 namespace GridKit
 {
   /*!
-   * @brief Declaration of a Hires Bus Component.
+   * @brief Hires Bus Component.
    *
    */
   template <class ScalarT, typename IdxT>
@@ -56,7 +56,7 @@ namespace GridKit
     {
     }
 
-    int allocate()
+    int allocate() final
     {
       CircuitComponent<ScalarT, IdxT>::allocate();
 
@@ -66,32 +66,30 @@ namespace GridKit
       return 0;
     }
 
-    int initialize()
+    int initialize() final
     {
       return 0;
     }
 
-    int tagDifferentiable()
+    int tagDifferentiable() final
     {
       return 0;
     }
 
-    int evaluateInternalResidual()
+    int evaluateInternalResidual() final
     {
       return 0;
     }
 
-    int evaluateExternalResidual()
+    int evaluateExternalResidual() final
     {
       *f_ext_[0] += -*yp_ext_[0] - *y_ext_[0];
       *f_ext_[1] += -*yp_ext_[1] - *y_ext_[1];
 
-      this->getResidual().setDataUpdated();
-
       return 0;
     }
 
-    int evaluateJacobian()
+    int evaluateJacobian() final
     {
       this->zeroJacMatrix();
 
@@ -104,22 +102,22 @@ namespace GridKit
       return 0;
     }
 
-    int evaluateIntegrand()
+    int evaluateIntegrand() final
     {
       return 0;
     }
 
-    int initializeAdjoint()
+    int initializeAdjoint() final
     {
       return 0;
     }
 
-    int evaluateAdjointResidual()
+    int evaluateAdjointResidual() final
     {
       return 0;
     }
 
-    int evaluateAdjointIntegrand()
+    int evaluateAdjointIntegrand() final
     {
       return 0;
     }
@@ -136,7 +134,7 @@ namespace GridKit
      * This represents a "noise" level close to zero for which pure relative
      * error cannot be used.
      */
-    int setAbsoluteTolerance(RealT rel_tol)
+    int setAbsoluteTolerance(RealT rel_tol) final
     {
       abs_tol_.setToConst(static_cast<ScalarT>(rel_tol));
       return 0;
